PR requested changes and refactors
